Switch PHS ALU kernel test to i32 data types
The schnitzel hardware uses 32-bit data width. Update addition.mlir,
test_software.mlir, and main.cpp to use i32/int32_t consistently.

+++ kernels/streamer_phs_alu/main.cpp
@@ -6,27 +6,27 @@
 #define DATA_LEN 16
 
 // MLIR-compiled kernel entry point
-extern "C" void _mlir_ciface_streamer_add(MemRefDescriptor<int64_t, 1> *A,
-                                          MemRefDescriptor<int64_t, 1> *B,
-                                          MemRefDescriptor<int64_t, 1> *O);
+extern "C" void _mlir_ciface_streamer_add(MemRefDescriptor<int32_t, 1> *A,
+                                          MemRefDescriptor<int32_t, 1> *B,
+                                          MemRefDescriptor<int32_t, 1> *O);
 
 // Test data in AXI/L3 memory (64-byte aligned for 512-bit AXI bus)
-alignas(64) int64_t A[DATA_LEN] = {1, 2,  3,  4,  5,  6,  7,  8,
-                                   9, 10, 11, 12, 13, 14, 15, 16};
-alignas(64) int64_t B[DATA_LEN] = {10, 20,  30,  40,  50,  60,  70,  80,
-                                   90, 100, 110, 120, 130, 140, 150, 160};
-alignas(64) int64_t O[DATA_LEN] = {0};
+alignas(64) int32_t A[DATA_LEN] = {1, 2,  3,  4,  5,  6,  7,  8,
+                                    9, 10, 11, 12, 13, 14, 15, 16};
+alignas(64) int32_t B[DATA_LEN] = {10, 20,  30,  40,  50,  60,  70,  80,
+                                    90, 100, 110, 120, 130, 140, 150, 160};
+alignas(64) int32_t O[DATA_LEN] = {0};
 
 // Expected: A + B
-int64_t G[DATA_LEN] = {11, 22,  33,  44,  55,  66,  77,  88,
-                       99, 110, 121, 132, 143, 154, 165, 176};
+int32_t G[DATA_LEN] = {11, 22,  33,  44,  55,  66,  77,  88,
+                        99, 110, 121, 132, 143, 154, 165, 176};
 
 int main() {
   int hart = hart_id();
 
-  MemRefDescriptor<int64_t, 1> memrefA = {A, A, 0, {DATA_LEN}, {1}};
-  MemRefDescriptor<int64_t, 1> memrefB = {B, B, 0, {DATA_LEN}, {1}};
-  MemRefDescriptor<int64_t, 1> memrefO = {O, O, 0, {DATA_LEN}, {1}};
+  MemRefDescriptor<int32_t, 1> memrefA = {A, A, 0, {DATA_LEN}, {1}};
+  MemRefDescriptor<int32_t, 1> memrefB = {B, B, 0, {DATA_LEN}, {1}};
+  MemRefDescriptor<int32_t, 1> memrefO = {O, O, 0, {DATA_LEN}, {1}};
 
   // The MLIR-compiled function handles DMA copies, accelerator CSR
   // programming, and synchronization internally.
@@ -39,8 +39,7 @@ int main() {
   if (hart == 1) {
     for (int i = 0; i < DATA_LEN; i++) {
       if (O[i] != G[i]) {
-        verbose_printf("  ERROR: O[%d] = %d, expected %d\n", i, (int)O[i],
-                       (int)G[i]);
+        verbose_printf("  ERROR: O[%d] = %d, expected %d\n", i, O[i], G[i]);
         err++;
       }
     }
